Some big changes to how useListenerState works
First things first, we need to be able to listen for changes on a given
field. That's been added. I also restructured the listeners state
accordingly, such that it's an object containing path-listener array
pairs, instead of a raw array of functions. Needs lots of testing but I
like the direction this is headed.