still seeing undefined flashing into state...
Ok, so I lied about taking a break. But now I really do need to step
away for a bit. The biggest issue is that undefined values are coming up
in the state, as in the state is changing to "undefined".

Need to unfrazzle my brain for a bit, unwind a little, then come back to
this later tonight or tomorrow